UPSTREAM: ARM: kvm: fix a bad BSYM() usage
BSYM() should only be used when refering to local symbols in the same
assembly file which are resolved by the assembler, and not for
linker-fixed up symbols.  The use of BSYM() with panic is incorrect as
the linker is involved in fixing up this relocation, and it knows
whether panic() is ARM or Thumb.

BUG=chromium:846515
TEST=build/boot on hana with USE=kvm_host

Change-Id: I686b1e6d907859b765bcb76f90af16a990c78a3a
Acked-by: Nicolas Pitre <nico@linaro.org>
Acked-by: Dave Martin <Dave.Martin@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
(cherry picked from commit 5890298a834c04aaa9b5fb576e5f2b77e79ab38d)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1088091