CHROMIUM: setlocalversion: check for .scmversion in output dir also
Look for .scmversion file in output directory in addition to the kernel
source tree. This is useful for builds where the output directory is not
the same as the kernel source directory. In such cases the .scmversion
file can be added/deleted in the output directory with no changes needed
in the kernel sources directories.

BUG=chrome-os-partner:40556
TEST=do kernel build for both cros_workon on and non-cros_workon
cases and verify that the kernel version is of the form 3.18.0 and
doesn't include the git hash in the kernel version.

Change-Id: I5d5a0259fe43960d428755205d1db7935b2a7d33
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/273276
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>

[rebase44(filbranden): Fixed script to cat the .scmversion found.]
Reviewed-on: https://chromium-review.googlesource.com/333874
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>
(cherry picked from commit 814f9c68230728c4d6dcb68f877c71d8cd9730b2)
Signed-off-by: Matt Atwood <matthew.s.atwood@intel.corp-partner.google.com>