UPSTREAM: arm/arm64: KVM: Fix ordering of timer/GIC on guest entry
As we now inject the timer interrupt when we're about to enter
the guest, it makes a lot more sense to make sure this happens
before the vgic code queues the pending interrupts.

Otherwise, we get the interrupt on the following exit, which is
not great for latency (and leads to all kind of bizarre issues
when using with active interrupts at the HW level).

BUG=chromium:846515
TEST=build/boot on hana with USE=kvm_host

Change-Id: I1f2ed9c288ee6d366de5a13447df7a800dd04dcf
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Reviewed-by: Alex Benne <alex.bennee@linaro.org>
Reviewed-by: Christoffer Dall <christoffer.dall@linaro.org>
(cherry picked from commit 9a99d050705318d1cb27979e1c810464347db9db)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1088116