UPSTREAM: ext4: try to initialize all groups we can in case of failure on ppc64
Currently on the machines with page size > block size when initializing
block group buddy cache we initialize it for all the block group bitmaps
in the page. However in the case of read error, checksum error, or if
a single bitmap is in any way corrupted we would fail to initialize all
of the bitmaps. This is problematic because we will not have access to
the other allocation groups even though those might be perfectly fine
and usable.

Fix this by reading all the bitmaps instead of error out on the first
problem and simply skip the bitmaps which were either not read properly,
or are not valid.

BUG=chromium:766786
TEST=build/boot caroline

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit bbdc322f2c600667e3d23dfadf8bbaad08f7edd3)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>

Change-Id: I5921e7d153ca7777ec23b6e1ddc432761075b013
Reviewed-on: https://chromium-review.googlesource.com/673752
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>