UPSTREAM: ext4: BUG_ON assertion repeated for inode1, not done for inode2
During a source code review of fs/ext4/extents.c I noted identical
consecutive lines. An assertion is repeated for inode1 and never done
for inode2. This is not in keeping with the rest of the code in the
ext4_swap_extents function and appears to be a bug.

Assert that the inode2 mutex is not locked.

BUG=chromium:766786
TEST=build/boot on caroline, xfstest

Change-Id: I1098aafb7b70b0f74aff3dfe2706970aaa554bce
Signed-off-by: David Moore <dmoorefo@gmail.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Eric Sandeen <sandeen@redhat.com>
(cherry picked from commit 8bc3b1e6e8fdc1c605c06c027d999b5cca434779)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/694893
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>