CHROMIUM: arm64: Cpu invariant scheduler load-tracking and capacity support
Provides the scheduler with a cpu scaling correction factor for more
accurate load-tracking and cpu capacity handling.

The Energy Model (EM) (in fact the capacity value of the last element
of the capacity states vector of the core (MC) level sched_group_energy
structure) is used as the source for this cpu scaling factor.

The cpu capacity value depends on the micro-architecture and the
maximum frequency of the cpu.

The maximum frequency part should not be confused with the frequency
invariant scheduler load-tracking support which deals with frequency
related scaling due to DFVS functionality.

BUG=chromium:580187
TEST=boot kevin/gru

Signed-off-by: Juri Lelli <juri.lelli@arm.com>
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>

Contains the following squashed commit due to build errors without it.
This is to make sure each incremental patch cleanly builds.
"arm64, topology: Updates to use DT bindings for EAS costing data"

Change-Id: I4581e5f5f1f7ec7ab3c8b58b7af476e3817f1b39
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/358221
Previous-Reviewed-on: https://chromium-review.googlesource.com/422989
(cherry picked from commit 1835ba770f55ac0a0a187175bc1a9d320a5a2013)
Reviewed-on: https://chromium-review.googlesource.com/423366
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>