ANDROID: sched/tune: add initial support for CGroups based boosting
To support task performance boosting, the usage of a single knob has the
advantage to be a simple solution, both from the implementation and the
usability standpoint.  However, on a real system it can be difficult to
identify a single value for the knob which fits the needs of multiple
different tasks. For example, some kernel threads and/or user-space
background services should be better managed the "standard" way while we
still want to be able to boost the performance of specific workloads.

In order to improve the flexibility of the task boosting mechanism this
patch is the first of a small series which extends the previous
implementation to introduce a "per task group" support.
This first patch introduces just the basic CGroups support, a new
"schedtune" CGroups controller is added which allows to configure
different boost value for different groups of tasks.
To keep the implementation simple but still effective for a boosting
strategy, the new controller:
  1. allows only a two layer hierarchy
  2. supports only a limited number of boost groups

A two layer hierarchy allows to place each task either:
  a) in the root control group
     thus being subject to a system-wide boosting value
  b) in a child of the root group
     thus being subject to the specific boost value defined by that
     "boost group"

The limited number of "boost groups" supported is mainly motivated by
the observation that in a real system it could be useful to have only
few classes of tasks which deserve different treatment.
For example, background vs foreground or interactive vs low-priority.
As an additional benefit, a limited number of boost groups allows also
to have a simpler implementation especially for the code required to
compute the boost value for CPUs which have runnable tasks belonging to
different boost groups.

cc: Tejun Heo <tj@kernel.org>
cc: Li Zefan <lizefan@huawei.com>
cc: Johannes Weiner <hannes@cmpxchg.org>
cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>

BUG=chrome-os-partner:58667
TEST=build/boot on kevin
CQ-DEPEND=CL:419990

Change-Id: I28a6089916e0e413e08a43ecbca7f68a6404bcf4
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
(cherry picked from commit 13001f47c9a610705219700af4636386b647e23
 from android-msm/android-4.4.y)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/420031
(cherry picked from commit 3f9f043cc68204dd7a54ac289d9a64b0bfca3217)
Reviewed-on: https://chromium-review.googlesource.com/423446