BACKPORT: ASoC: rt5645: Update dapm pins when the card is not instantiated yet
This makes sure the dapm state is consistent when the card is
instantiated.

However, if the card is not instantiated yet, we still update the
registers manually in the "jack in" case, so that we can immediately
report if a mic is present or not.

Disabling "Mic Det Power" after detection, and on jack out, can
wait until the card gets instantiated.

Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit b14c917409f47dbfdff93e69effd6740f6a809f7,
 https://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git)
Signed-off-by: John Lin <john.lin@realtek.com>

Conflicts:
	sound/soc/codecs/rt5645.c - context fix

BUG=chrome-os-partner:49287
TEST=normal boot and verify headset detection

Change-Id: Ifbb122498ce0a0ed58873c462d3becc946c7cc9d
Reviewed-on: https://chromium-review.googlesource.com/322088
Commit-Ready: John Lin <john.lin@realtek.com>
Tested-by: John Lin <john.lin@realtek.com>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: John Lin <john.lin@realtek.com>