UPSTREAM: drm/i915: Keep the mm.bound_list in rough LRU order
When we shrink our working sets, we want to avoid stealing pages from
objects that likely to be reused in the near future. We first look at
inactive objects before processing active objects - but what about a
recently active object that is about to be used again. That object's
position in the bound_list is ordered by the time of binding, not the
time of last use, so the most recently used inactive object could well
be at the head of the shrink list. To compensate, give the object a bump
to MRU when it becomes inactive (thus transitioning to the end of the
first pass in shrink lists). Conversely, bumping on inactive makes
bumping on active useless, since when we do have to reap from the active
working set, everything is going to become inactive very quickly and the
order pretty much random - just hope for the best at that point, as once
we start stalling on active objects, we can hope that the rebinding
neatly orders vital objects.

Suggested-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
[danvet: Resolve merge conflict.]
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 6c24695988774d56ea8d6a47ffdaf0fd32072488)

BUG=b:37163270
TEST=test use cases described on bug

Change-Id: I96b1133a288b9fafbeb06344256b873d7faedbd1
Signed-off-by: Lee, Shawn C <shawn.c.lee@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/860323
Commit-Ready: Daniel Charles <daniel.charles@intel.com>
Tested-by: Carlos Santa <carlos.santa@intel.com>
Tested-by: Daniel Charles <daniel.charles@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Carlos Santa <carlos.santa@intel.com>
Reviewed-by: Shawn C Lee <shawn.c.lee@intel.com>
(cherry picked from commit dcc7ab290be03abe27b5313d9d947d25678c4a63)
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/897086