UPSTREAM: HID: sony: Fix DS4 controller reporting rate issues
This commit removes the cap on the DualShock 4 controller reporting rate when
connected using Bluetooth. The previous value of '0xB0' capped the rate to only
20.83 Hz which many userspace utilities mistook as a sign of a bad signal.
Since a 'B' and an '8' can look similar it's possible that someone mistook the
one for another.  The new value of '0x80' enables the full 1000 Hz peak
reporting rate that the controller is capable of.

Frank adds: "Back when the original code was written the purpose of that value
was unknown and 0xB0 seemed to work so that's what ended up being used.  Now
that we know what it actually does and that 0x80 is a better choice I support
this patch."

[jkosina@suse.com: update changelog]
Signed-off-by: Rostislav Pehlivanov <atomnuker@gmail.com>
Acked-by: Frank Praznik <frank.praznik@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.com>
(cherry picked from commit 824deff87f5b1a2f32b9c3164191b0f5b22276b8)

BUG=b:139939892, chromium:972297
TEST=DualShock v4 via Bluetooth works on https://html5gamepad.com/

Change-Id: I5eab53f3fc753db8532205bf849a2d048c244372
Signed-off-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1862759
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>