CHROMIUM: iwl7000: correctly size command string arrays
The command string array is assumed to be indexable by a full u8, so it
must have 256 entries (0-255), not just 255. A recent firmware change
(apparently) started using the command 0xff for a notification to the
host, causing the driver to crash in debug message/tracing code.

For safety, fix this for all the opmodes, even if only the recently
changed firmware is able to trigger this.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 219e6c80746429931596468e34c672341e07c1bb