UPSTREAM: crypto: skcipher - Add top-level skcipher interface
This patch introduces the crypto skcipher interface which aims
to replace both blkcipher and ablkcipher.

It's very similar to the existing ablkcipher interface.  The
main difference is the removal of the givcrypt interface.  In
order to make the transition easier for blkcipher users, there
is a helper SKCIPHER_REQUEST_ON_STACK which can be used to place
a request on the stack for synchronous transforms.

BUG=chromium:704196
TEST=Deploy v5.44 BlueZ system daemon with v4.9 BlueZ kernel on Oak and
perform automated tests (suite:bluetooth) and manual tests against mouses,
keyboards and headphones.

Change-Id: I1bd30d9bdb3f14df990bcd90f892ea717cfc80c1
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Miao-chen Chou <mcchou@chromium.org>
(cherry picked from commit 7a7ffe65c8c5fbf272b132d8980b2511d5e5fc98)
Reviewed-on: https://chromium-review.googlesource.com/474084
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>