UPSTREAM: drm/i915: Refactor the hpd irq handling functions
A lot of the hpd irq handling is duplicated code, so refactor it a bit
by observing that in several places the only difference is the hpd[]
array. So pull the code to a few functions and pass in the hpd[] array
from the caller. Another option would be to determine the correct array
to use within the functions themselves, but somehow passing it in felt
nicer.

Further code reduction could be achieved by passing in the hotplug
register offset, and the long pulse detection function pointer. But that
didn't feel as good for some reason, so I left it at the middle ground.

Signed-off-by: Ville Syrjl <ville.syrjala@linux.intel.com>
Reviewed-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 40e5641086f6f64acb13ff6893cf82c4fb30a6d7)
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>

BUG=chrome-os-partner:53839
TEST=External monitors connected via MST hub or daisy-chain over DP should come up.

Change-Id: I746fac815ff9c63dd013622555dd87ed73db85cb
Reviewed-on: https://chromium-review.googlesource.com/462109
Commit-Ready: Brian J Lovin <brian.j.lovin@intel.com>
Tested-by: Chenglei Ren <chenglei.ren@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>