BACKPORT: flow_dissector: Jump to exit code in __skb_flow_dissect
Instead of returning immediately (on a parsing failure for instance) we
jump to cleanup code. This always sets protocol values in key_control
(even on a failure there is still valid information in the key_tags that
was set before the problem was hit).

BUG=chromium:763645
TEST=Build and run

Change-Id: If00aa1704cc260f47d541c974c0fbf31e859d89c
Signed-off-by: Tom Herbert <tom@herbertland.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
[backport: Various code flow changes]
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit a6e544b0a88b)
Reviewed-on: https://chromium-review.googlesource.com/658997
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Conflicts:
	net/core/flow_dissector.c
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/663526