CHROMIUM: iwl7000: mvm: remove pointless cfg_phy_cnt length check
Since the driver can never configure the data here, this field
will always be reported as 0 by the firmware. Even if this was
not the case, however, it wouldn't matter since the extra data
would be added beyond the end of the phy_info structure we use
in the driver, so wouldn't harm anything in this code either.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: d3fb0eeb43bb29903be7fe0e56b0c24152f48877