CHROMIUM: iio: core: Add double tap gesture
Add double tap as possible gesture.
This is an inteface change: however, the sysfs entry is based on string,
so if other gestures are added on the trunk in the meantime, we will
still be able to merge this change.

BUG=chrome-os-partner:44754
TEST=compile. Check with cros_iio_activity driver.

Change-Id: I7700b02a8172b59975f8c1eda3e4e3b6d1b93508
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/297838
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Conflicts:
	drivers/iio/industrialio-core.c
	include/linux/iio/types.h
[rebase44(groeck): Resolved conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Conflicts:
	include/linux/iio/types.h
[rebase49(groeck): Resolved conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit c0c554e86552f639be8c7df5ccba85de88113a4a)
Signed-off-by: Matt Atwood <matthew.s.atwood@intel.corp-partner.google.com>