UPSTREAM: net: stmmac: Use msleep rather then udelay for reset delay
The reset delays used for stmmac are in the order of 10ms to 1 second,
which is far too long for udelay usage, so switch to using msleep.

Practically this fixes the PHY not being reliably detected in some cases
as udelay wouldn't actually delay for long enough to let the phy
reliably be reset.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 892aa01df2ad67237f213c8f9d9b491e908aa910)

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:53191
TEST=manual: with series, on Tiger, suspend the machine and wake it up
from keyboard and using a WoL magic packet, and check the on-board
networking is still working.

Change-Id: I7c0f111600d0286e09ceeae939503615633dbee6
Reviewed-on: https://chromium-review.googlesource.com/348851
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>