BACKPORT: x86/nmi/64: Fix a paravirt stack-clobbering bug in the NMI code
[ Upstream commit 83c133cf11fb0e68a51681447e372489f052d40e ]

The NMI entry code that switches to the normal kernel stack needs to
be very careful not to clobber any extra stack slots on the NMI
stack.  The code is fine under the assumption that SWAPGS is just a
normal instruction, but that assumption isn't really true.  Use
SWAPGS_UNSAFE_STACK instead.

This is part of a fix for some random crashes that Sasha saw.

BUG=chromium:842936
TEST=build/boot on caroline with USE=kvm_host

Change-Id: Iee37483403d7f78c59328793815786465b2fe3f9
Fixes: 9b6e6a8334d5 ("x86/nmi/64: Switch stacks on userspace NMI entry")
Reported-and-tested-by: Sasha Levin <sasha.levin@oracle.com>
Signed-off-by: Andy Lutomirski <luto@kernel.org>
Cc: stable@vger.kernel.org
Link: http://lkml.kernel.org/r/974bc40edffdb5c2950a5c4977f821a446b76178.1442791737.git.luto@kernel.org
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Sasha Levin <sasha.levin@oracle.com>
(cherry picked from commit a70046a8f849fb1a98c5aee0c3504ddf61461eab
 in https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git/)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
[SR: context conflict with 565ec1909b081e]
 Conflicts:
	arch/x86/kernel/entry_64.S
Reviewed-on: https://chromium-review.googlesource.com/1068415
Reviewed-by: Dylan Reid <dgreid@chromium.org>