UPSTREAM: drm/i915: Report context GTT size
Since the beginning we have conflated the size of the global GTT with
that of the per-process context sizes. In recent times (gen8+), those
are no longer the same where the global GTT is limited to 2/4GiB but the
per-process GTT may be anything up to 256TiB. Userspace knows nothing of
this discrepancy and outside of one or two hacks, uses the getaperture
ioctl to determine the maximum size it can use. Let's leave that as
reporting the global GTT and use the context reporting method to
describe the per-process value (which naturally fallsback to reporting
the aliasing or global on older platforms, so userspace can always use
this method where available).

Testcase: igt/gem_userptr_blits/minor-normal-sync
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=90065

BUG=b:35648344
TEST=Run Hangouts, Youtube, and Webgl Aquarium together.

Change-Id: I4bc7a01924e7854d4bcc2213a3bbd34d0a4a7cf4
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit fa8848f27895bd19e16aed77868f464be24034e6)
Signed-off-by: Yu Kang Ku <yu.kang.ku@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/505705
Tested-by: Daniel Charles <daniel.charles@intel.com>
Tested-by: Michael Cheng <michael.cheng@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>