CHROMIUM: PM / sleep: try to runtime suspend for direct complete
This tries to runtime suspend devices that are still active for direct
complete. This is for cases such as autosuspend delays which leaves devices
able to runtime suspend but still active. It's beneficial in this case to
runtime suspend the device to take advantage of direct complete when
possible.

BUG=chrome-os-partner:46491
TEST=suspend_stress_test on glados with usb drive plugged in

Change-Id: Ia935eee214ef6cd59097f9c60c39dbb89a0bb27e
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/307565
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
(cherry picked from commit 6c650102b009f3148aefef047856f0e3c64820d1)
Signed-off-by: Matt Atwood <matthew.s.atwood@intel.corp-partner.google.com>