BACKPORT: HID: sony: Save and restore the controller state on suspend and resume
On hardware which provides standby power for charging devices the state of the
LEDs and force-feedback on controllers can persist even when the system is in
standby.  Additionally, the state of the controllers on resume may be different
from the state they were in at the time when they were suspended (ie. LEDs are
cleared on resume).

This implements the suspend and resume callbacks which saves and clears the
state of the LEDs on suspend and restores them on resume.  Force-feedback is
stopped on suspend but not automatically restored on resume until a new event is
received to avoid potentially damaging hardware.

USB Sixaxis and navigation controllers must be reinitialized when the hardware
is reset on resume or they won't send any input reports.

Signed-off-by: Frank Praznik <frank.praznik@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit decd946c99f6b3826bda0bfd5d1b2ddd56ef6b54)

Conflicts:
   drivers/hid/hid-sony.c

BUG=b:139939892, chromium:972297
TEST=DualShock v4 via Bluetooth works on https://html5gamepad.com/

Change-Id: I63928c4c58c5b5eb3b0c1327cb33e8ae05c6c4ca
Signed-off-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1800243
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>